moved cron job into Heroku scheduler task
